chore: move ReplyCard
